requirements.txt
Django~=5.1.5
pillow==11.1.0
django-tailwind==3.8.0
django-browser-reload==1.17.0      # สำหรับ Dev, reload หน้าอัตโนมัติ
django-tinymce==4.1.0
#mysqlclient==2.2.7
requests==2.32.3
django-environ==0.12.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
gunicorn==23.0.0